all deficiencies fixed by mentors feedbacks

+++ domain/src/main/java/com/epam/esm/dto/response/MostUsedTagResponse.java
@@ -15,5 +15,5 @@ import java.util.List;
 public class MostUsedTagResponse {
 
     List<TagGetResponse> tags;
-    BigInteger count;
+    Integer count;
 }
